Update linked list (task4)
Code changes in main(). Long code replaced by concise one. Also, before was a problem if the first line is '.' a list will be crafted. That was a mistake of conditions loop organization. Now, first of all, the program checks whether the first line is '.' or not. If it is not we check whether the string is NULL or not. 

Also, some stylistic changes are required, for example, pop and push functions names are not apropos.